requirements_test.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-homeassistant-custom-component>=0.13.190
